refactor: 'modules' dir was deprecated
