+++ pypedro/helper/db/mongodb.py
@@ -4,7 +4,7 @@ from abc import ABCMeta
 
 from mongoengine import connect as connect_db, disconnect_all
 
-from pypedro.modules.helper.db.interfaces import IConfig
+from pypedro.helper.db.interfaces import IConfig
 
 
 class Config(IConfig, metaclass=ABCMeta):
